+++ CHANGELOG.md
@@ -29,12 +29,14 @@
 - docker hub にプッシュするスクリプト。
 
 ## v2.9.899 / 2024-05-01
-- resources/typing_ex/config.edn:
+- post "/exam"
+  - resources/typing_ex/config.edn:
   [post "/exam"]
-- src/typing_ex/handler/core.clj:
+  - src/typing_ex/handler/core.clj:
   (defmethod exam!)
-- src/typing_ex/typing.cljs:
+  - src/typing_ex/typing.cljs:
   (defn- exam-point! [login count pt])
+- get "/exam/:login/:ct"
 
 ## v2.8.893 / 2024-04-29
 - display end `datetime` in `/rc`.